fix 32 bit

+++ ChunkCodecTests/src/ChunkCodecTests.jl
@@ -72,7 +72,7 @@ function test_codec(c::Codec, e::EncodeOptions, d::DecodeOptions; trials=100)
             @test isnothing(try_encode!(e, zeros(UInt8, length(encoded)-1), data))
         end
         local ds = try_find_decoded_size(d, encoded)
-        @test ds isa Union{Nothing, Int}
+        @test ds isa Union{Nothing, Int64}
         if !isnothing(ds)
             @test ds === s
         end
@@ -93,21 +93,21 @@ function test_codec(c::Codec, e::EncodeOptions, d::DecodeOptions; trials=100)
             dst = zeros(UInt8, s - 1)
             @test_throws(
                 ArgumentError("`max_size`: $(-1) must be at least `length(dst)`: $(s-1)"),
-                try_resize_decode!(d, dst, encoded; max_size=-1)
+                try_resize_decode!(d, dst, encoded; max_size=Int64(-1))
             )
             dst = zeros(UInt8, s - 1)
             @test try_resize_decode!(d, dst, encoded; max_size=s) == s
             @test length(dst) == s
             @test dst == data
             dst = UInt8[]
-            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=0))
+            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=Int64(0)))
         end
         if s > 1
             dst = UInt8[]
-            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=1))
+            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=Int64(1)))
             dst = UInt8[0x01]
-            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=1))
-            @test_throws DecodedSizeError(1, try_find_decoded_size(d, encoded)) decode(d, encoded; max_size=1)
+            @test isnothing(try_resize_decode!(d, dst, encoded; max_size=Int64(1)))
+            @test_throws DecodedSizeError(1, try_find_decoded_size(d, encoded)) decode(d, encoded; max_size=Int64(1))
         end
         dst_buffer = zeros(UInt8, s + 2)
         dst = view(dst_buffer, 1:s+1)
